backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
openai==1.2.0
python-dotenv==1.0.0
PyPDF2==3.0.1
docx2txt==0.8
langchain==0.0.335